orders_djg/requirements.txt
channels==4.1.0
Django==5.0.1
Pillow==10.3.0
PyJWT==2.8.0
redis==5.0.4
